Separate synchronous configuration manager
