--- src/Microsoft.IdentityModel.Protocols.OpenIdConnect/PublicAPI.Unshipped.txt
@@ -1,1 +1,0 @@
-static Microsoft.IdentityModel.Protocols.OpenIdConnect.OpenIdConnectConfigurationRetriever.GetSync(string address, Microsoft.IdentityModel.Protocols.IDocumentRetrieverSync retriever, System.Threading.CancellationToken cancel) -> Microsoft.IdentityModel.Protocols.OpenIdConnect.OpenIdConnectConfiguration
