--- src/Microsoft.IdentityModel.Protocols.WsFederation/PublicAPI.Unshipped.txt
@@ -1,2 +1,0 @@
-Microsoft.IdentityModel.Protocols.WsFederation.WsFederationConfigurationRetriever.GetConfigurationSync(string address, Microsoft.IdentityModel.Protocols.IDocumentRetrieverSync retriever, System.Threading.CancellationToken cancel) -> Microsoft.IdentityModel.Protocols.WsFederation.WsFederationConfiguration
-static Microsoft.IdentityModel.Protocols.WsFederation.WsFederationConfigurationRetriever.GetSync(string address, Microsoft.IdentityModel.Protocols.IDocumentRetrieverSync retriever, System.Threading.CancellationToken cancel) -> Microsoft.IdentityModel.Protocols.WsFederation.WsFederationConfiguration
