+++ src/Microsoft.IdentityModel.Protocols/InternalAPI.Unshipped.txt
@@ -14,5 +14,3 @@ Microsoft.IdentityModel.Protocols.IDocumentRetrieverSync.GetDocument(string addr
 const Microsoft.IdentityModel.Protocols.LogMessages.IDX20811 = "IDX20811: Failed to retrieve configuration from event handler. MetadataAddress: '{0}', Exception: '{1}'." -> string
 const Microsoft.IdentityModel.Protocols.LogMessages.IDX20812 = "IDX20812: Failed to validate configuration retrieved via event handler. Exception: '{0}'." -> string
 const Microsoft.IdentityModel.Protocols.LogMessages.IDX20813 = "IDX20813: ConfigurationEventHandler implementation threw an exception on AfterUpdateAsync call for metadata address '{0}'. Exception: '{1}'." -> string
-const Microsoft.IdentityModel.Protocols.LogMessages.IDX20814 = "IDX20814: The synchronous configuration retrieval path is not available on this ConfigurationManager. It requires a configuration retriever that implements 'IConfigurationRetrieverSync<T>' and a document retriever that implements 'IDocumentRetrieverSync'. Create the manager with ConfigurationManager<T>.CreateSync(...), or provide retrievers that implement the synchronous interfaces." -> string
-const Microsoft.IdentityModel.Protocols.LogMessages.IDX20815 = "IDX20815: The asynchronous configuration retrieval path is not available on this ConfigurationManager. It requires a configuration retriever that implements 'IConfigurationRetriever<T>' and a document retriever that implements 'IDocumentRetriever'." -> string
