+++ src/Microsoft.IdentityModel.Protocols/PublicAPI.Unshipped.txt
@@ -1,11 +1,3 @@
 Microsoft.IdentityModel.Protocols.FileDocumentRetriever.GetDocument(string address, System.Threading.CancellationToken cancel) -> string
-Microsoft.IdentityModel.Protocols.ConfigurationManager<T>.GetConfigurationSync() -> T
-virtual Microsoft.IdentityModel.Protocols.ConfigurationManager<T>.GetConfigurationSync(System.Threading.CancellationToken cancel) -> T
-override Microsoft.IdentityModel.Protocols.ConfigurationManager<T>.GetBaseConfigurationSync(System.Threading.CancellationToken cancel) -> Microsoft.IdentityModel.Tokens.BaseConfiguration
 Microsoft.IdentityModel.Protocols.StaticConfigurationManager<T>.GetConfigurationSync(System.Threading.CancellationToken cancel) -> T
 override Microsoft.IdentityModel.Protocols.StaticConfigurationManager<T>.GetBaseConfigurationSync(System.Threading.CancellationToken cancel) -> Microsoft.IdentityModel.Tokens.BaseConfiguration
-static Microsoft.IdentityModel.Protocols.ConfigurationManager<T>.CreateSync(string metadataAddress, Microsoft.IdentityModel.Protocols.IConfigurationRetrieverSync<T> configRetriever, Microsoft.IdentityModel.Protocols.IDocumentRetrieverSync docRetriever) -> Microsoft.IdentityModel.Protocols.ConfigurationManager<T>
-static Microsoft.IdentityModel.Protocols.ConfigurationManager<T>.CreateSync(string metadataAddress, Microsoft.IdentityModel.Protocols.IConfigurationRetrieverSync<T> configRetriever, Microsoft.IdentityModel.Protocols.IDocumentRetrieverSync docRetriever, Microsoft.IdentityModel.Protocols.Configuration.LastKnownGoodConfigurationCacheOptions lkgCacheOptions) -> Microsoft.IdentityModel.Protocols.ConfigurationManager<T>
-static Microsoft.IdentityModel.Protocols.ConfigurationManager<T>.CreateSync(string metadataAddress, Microsoft.IdentityModel.Protocols.IConfigurationRetrieverSync<T> configRetriever, Microsoft.IdentityModel.Protocols.IDocumentRetrieverSync docRetriever, Microsoft.IdentityModel.Protocols.IConfigurationValidator<T> configValidator) -> Microsoft.IdentityModel.Protocols.ConfigurationManager<T>
-static Microsoft.IdentityModel.Protocols.ConfigurationManager<T>.CreateSync(string metadataAddress, Microsoft.IdentityModel.Protocols.IConfigurationRetrieverSync<T> configRetriever, Microsoft.IdentityModel.Protocols.IDocumentRetrieverSync docRetriever, Microsoft.IdentityModel.Protocols.IConfigurationValidator<T> configValidator, Microsoft.IdentityModel.Protocols.Configuration.LastKnownGoodConfigurationCacheOptions lkgCacheOptions) -> Microsoft.IdentityModel.Protocols.ConfigurationManager<T>
-static Microsoft.IdentityModel.Protocols.ConfigurationManager<T>.CreateSync(string metadataAddress, Microsoft.IdentityModel.Protocols.IConfigurationRetrieverSync<T> configRetriever, Microsoft.IdentityModel.Protocols.IDocumentRetrieverSync docRetriever, Microsoft.IdentityModel.Protocols.IConfigurationValidator<T> configValidator, Microsoft.IdentityModel.Protocols.Configuration.LastKnownGoodConfigurationCacheOptions lkgCacheOptions, Microsoft.IdentityModel.Protocols.Configuration.IConfigurationEventHandlerSync<T> configurationEventHandler) -> Microsoft.IdentityModel.Protocols.ConfigurationManager<T>
